+++ gridsome/lib/graphql/__tests__/nodes.pagination.spec.js
@@ -2,8 +2,9 @@ const App = require('../../app/App')
 const { BOOTSTRAP_PAGES } = require('../../utils/constants')
 
 test('should return all nodes', async () => {
-  const results = await graphql()
+  const results = await graphql(100)
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(100)
   expect(results.data.allPost.pageInfo).toMatchObject({
@@ -18,9 +19,28 @@ test('should return all nodes', async () => {
   })
 })
 
-test('return limited nodes', async () => {
-  const results = await graphql('limit: 10')
+test('should return empty results', async () => {
+  const results = await graphql(0)
 
+  expect(results.errors).toBeUndefined()
+  expect(results.data.allPost.totalCount).toEqual(0)
+  expect(results.data.allPost.edges).toHaveLength(0)
+  expect(results.data.allPost.pageInfo).toMatchObject({
+    perPage: 0,
+    totalPages: 1,
+    currentPage: 1,
+    totalItems: 0,
+    hasPreviousPage: false,
+    hasNextPage: false,
+    isFirst: true,
+    isLast: true
+  })
+})
+
+test('return limited nodes', async () => {
+  const results = await graphql(100, 'limit: 10')
+
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(10)
   expect(results.data.allPost.edges[0].node.id).toEqual('1')
@@ -38,8 +58,9 @@ test('return limited nodes', async () => {
 })
 
 test('skip nodes', async () => {
-  const results = await graphql('skip: 95')
+  const results = await graphql(100, 'skip: 95')
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(5)
   expect(results.data.allPost.edges[0].node.id).toEqual('96')
@@ -57,8 +78,9 @@ test('skip nodes', async () => {
 })
 
 test('limit and skip', async () => {
-  const results = await graphql('limit: 10', 'skip: 10')
+  const results = await graphql(100, 'limit: 10', 'skip: 10')
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(10)
   expect(results.data.allPost.edges[0].node.id).toEqual('11')
@@ -76,8 +98,9 @@ test('limit and skip', async () => {
 })
 
 test('limit and skip is more than total nodes', async () => {
-  const results = await graphql('limit: 10', 'skip: 95')
+  const results = await graphql(100, 'limit: 10', 'skip: 95')
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(5)
   expect(results.data.allPost.edges[0].node.id).toEqual('96')
@@ -95,8 +118,9 @@ test('limit and skip is more than total nodes', async () => {
 })
 
 test('limit results with perPage argument', async () => {
-  const results = await graphql('perPage: 10')
+  const results = await graphql(100, 'perPage: 10')
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(10)
   expect(results.data.allPost.pageInfo).toMatchObject({
@@ -112,8 +136,9 @@ test('limit results with perPage argument', async () => {
 })
 
 test('return specific page', async () => {
-  const results = await graphql('page: 2')
+  const results = await graphql(100, 'page: 2')
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(25)
   expect(results.data.allPost.edges[0].node.id).toEqual('26')
@@ -131,8 +156,9 @@ test('return specific page', async () => {
 })
 
 test('return specific page within limit', async () => {
-  const results = await graphql('page: 1', 'limit: 50')
+  const results = await graphql(100, 'page: 1', 'limit: 50')
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(25)
   expect(results.data.allPost.edges[0].node.id).toEqual('1')
@@ -150,8 +176,9 @@ test('return specific page within limit', async () => {
 })
 
 test('return specific page within limit with custom perPage', async () => {
-  const results = await graphql('page: 5', 'perPage: 3', 'limit: 50')
+  const results = await graphql(100, 'page: 5', 'perPage: 3', 'limit: 50')
 
+  expect(results.errors).toBeUndefined()
   expect(results.data.allPost.totalCount).toEqual(100)
   expect(results.data.allPost.edges).toHaveLength(3)
   expect(results.data.allPost.edges[0].node.id).toEqual('13')
@@ -168,7 +195,7 @@ test('return specific page within limit with custom perPage', async () => {
   })
 })
 
-async function graphql (...args) {
+async function graphql (count, ...args) {
   const defaultSort = 'sort: [{ by: "order", order: ASC }]'
   const argsArr = [...args, defaultSort]
 
@@ -177,9 +204,9 @@ async function graphql (...args) {
       plugins: [
         function plugin (api) {
           api.loadSource(store => {
-            const posts = store.addContentType('Post')
+            const posts = store.addCollection('Post')
 
-            for (let i = 1; i <= 100; i++) {
+            for (let i = 1; i <= count; i++) {
               posts.addNode({ id: String(i), order: i })
             }
           })