+++ gridsome/lib/graphql/nodes/utils.js
@@ -1,4 +1,3 @@
-const { safeKey } = require('../../utils')
 const { PER_PAGE } = require('../../utils/constants')
 
 exports.applyChainArgs = function (chain, args = {}, sort = []) {
@@ -7,10 +6,6 @@ exports.applyChainArgs = function (chain, args = {}, sort = []) {
   if (args.limit) chain = chain.limit(args.limit)
 
   return chain
-}
-
-exports.createBelongsToKey = function (node) {
-  return `belongsTo.${node.typeName}.${safeKey(node.id)}`
 }
 
 exports.createPagedNodeEdges = function (chain, args = {}, sort = []) {
@@ -48,7 +43,7 @@ exports.createPagedNodeEdges = function (chain, args = {}, sort = []) {
 
   const nodes = chain.data()
   const currentPage = page
-  const totalPages = Math.max(Math.ceil(totalItemsCount / perPage), 1)
+  const totalPages = Math.max(Math.ceil(totalItemsCount / perPage) || 1, 1)
   const hasPreviousPage = page > 1
   const hasNextPage = page < totalPages