+++ packages/transformer-remark/lib/utils.js
@@ -19,29 +19,26 @@ exports.createFile = function (node) {
   })
 }
 
-exports.normalizePlugins = function (arr = []) {
-  const normalize = entry => {
-    return typeof entry === 'string'
-      ? require(entry)
-      : entry
-  }
-
-  return arr.map(entry => {
-    return Array.isArray(entry)
-      ? [normalize(entry[0]), entry[1] || {}]
-      : [normalize(entry), {}]
-  })
-}
-
-exports.createPlugins = function (options, userPlugins) {
+exports.createPlugins = function (options, localOptions) {
+  const userPlugins = (options.plugins || []).concat(localOptions.plugins || [])
   const plugins = []
 
   if (options.useBuiltIns === false) {
-    return exports.normalizePlugins(userPlugins || [])
+    return normalizePlugins(userPlugins || [])
   }
 
-  plugins.push(require('./plugins/file'))
-  plugins.push(require('./plugins/image'))
+  if (options.processFiles !== false) {
+    plugins.push(require('./plugins/file'))
+  }
+
+  if (options.processImages !== false) {
+    plugins.push([require('./plugins/image'), {
+      blur: options.imageBlurRatio,
+      quality: options.imageQuality,
+      background: options.imageBackground,
+      immediate: options.lazyLoadImages === false ? true : undefined
+    }])
+  }
 
   if (options.slug !== false) {
     plugins.push('remark-slug')
@@ -80,7 +77,7 @@ exports.createPlugins = function (options, userPlugins) {
 
   plugins.push(...userPlugins)
 
-  return exports.normalizePlugins(plugins)
+  return normalizePlugins(plugins)
 }
 
 exports.findHeadings = function (ast) {
@@ -104,4 +101,18 @@ exports.findHeadings = function (ast) {
   })
 
   return headings
+}
+
+function normalizePlugins (arr = []) {
+  const normalize = entry => {
+    return typeof entry === 'string'
+      ? require(entry)
+      : entry
+  }
+
+  return arr.map(entry => {
+    return Array.isArray(entry)
+      ? [normalize(entry[0]), entry[1] || {}]
+      : [normalize(entry), {}]
+  })
 }