+++ packages/source-wordpress/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/gridsome/gridsome/tree/master/packages/source-wordpress/compare/@gridsome/source-wordpress@0.4.0...@gridsome/source-wordpress@0.4.1) (2019-09-17)
+
+
+### Bug Fixes
+
+* **wordpress:** don’t add default routes ([6e7134f](https://github.com/gridsome/gridsome/tree/master/packages/source-wordpress/commit/6e7134f))
+
+
+
+
+
+# [0.4.0](https://github.com/gridsome/gridsome/tree/master/packages/source-wordpress/compare/@gridsome/source-wordpress@0.3.0...@gridsome/source-wordpress@0.4.0) (2019-09-11)
+
+
+### Features
+
+* **wordpress:** basic gridsome v0.7 support ([8433c9a](https://github.com/gridsome/gridsome/tree/master/packages/source-wordpress/commit/8433c9a))
+
+
+
+
+
+# [0.3.0](https://github.com/gridsome/gridsome/tree/master/packages/source-wordpress/compare/@gridsome/source-wordpress@0.2.2...@gridsome/source-wordpress@0.3.0) (2019-05-10)
+
+
+### Features
+
+* **wordpress:** camelcase fields and use latest apis ([0c431cf](https://github.com/gridsome/gridsome/tree/master/packages/source-wordpress/commit/0c431cf))
+
+
+
+
+
 ## [0.2.2](https://github.com/gridsome/gridsome/tree/master/packages/source-wordpress/compare/@gridsome/source-wordpress@0.2.1...@gridsome/source-wordpress@0.2.2) (2019-04-18)